// src/poker/types.ts
export interface Card {
  figure:
    | "A"
    | "2"
    | "3"
    | "4"
    | "5"
    | "6"
    | "7"
    | "8"
    | "9"
    | "10"
    | "J"
    | "Q"
    | "K";
  suite: "hearts" | "diamond" | "spades" | "club";
  value: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13 | 14;
}

export interface CardSuitCounterMap {
  [key: string]: number;
}